♻️ homeのコードを機能ごとに分割
